Refactor FeeProvider
This commit changes the fee provider to estimate the fee to pay based on the
exchange rate. Since Bitcoin Cash blocks are not full pretty much any fee
will get mined. This commit changes it to target a fee-per-byte such that a median
size transaction (226 bytes) would cost 1 USD cent for economic, 5 cents for
normal, and 10 cents for priority.